requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion Models
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0

# ControlNet
controlnet-aux>=0.0.7

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# Depth Estimation
timm>=0.9.0

# Utilities
safetensors>=0.4.0
huggingface-hub>=0.19.0
